refactor: isolate formatting tests
